packages/cli/pyproject.toml
[project]
name = "cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "common",
    "mistune>=3.2.0",
    "pypdf>=6.7.1",
    "pdfminer-six>=20260107",
]

[project.scripts]
init = "init_scripts:main"

[tool.uv.sources]
common = { workspace = true }
